refactor: rename api-util
